+++ src/test/symbols/replacements.test.js
@@ -3,8 +3,6 @@ import assert from 'assert';
 
 describe('Replace following symbols\n', () => {
 	let testCase = {
-		"(C) (C)": "© ©",
-		"(c) (c)": "© ©",
 		"(P) (P)": "℗ ℗",
 		"(p) (p)": "℗ ℗",
 		"(R) (R)": "® ®",